+++ dwds/test/chrome_proxy_service_test.dart
@@ -8,10 +8,9 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:dwds/src/connections/debug_connection.dart';
+import 'package:dwds/src/loaders/strategy.dart';
 import 'package:dwds/src/services/chrome_proxy_service.dart';
 import 'package:dwds/src/utilities/dart_uri.dart';
-import 'package:dwds/src/utilities/shared.dart';
-
 import 'package:http/http.dart' as http;
 import 'package:path/path.dart' as path;
 import 'package:pedantic/pedantic.dart';
@@ -447,7 +446,7 @@ void main() {
       Future<RemoteObject> createList() {
         var expr = '''
           (function () {
-            const sdk = $loadModule("dart_sdk");
+            const sdk = ${globalLoadStrategy.loadModuleSnippet}("dart_sdk");
             const list = sdk.dart.dsend(sdk.core.List,"filled", [1001, 5]);
             list[4] = 100;
             return list;
@@ -459,10 +458,10 @@ void main() {
       Future<RemoteObject> createMap() {
         var expr = '''
           (function () {
-            const sdk = $loadModule("dart_sdk");
+            const sdk = ${globalLoadStrategy.loadModuleSnippet}("dart_sdk");
             const iterable = sdk.dart.dsend(sdk.core.Iterable, "generate", [1001]);
             const list1 = sdk.dart.dsend(iterable, "toList", []);
-            const reversed = sdk.dart.dload(list1, "reversed"); 
+            const reversed = sdk.dart.dload(list1, "reversed");
             const list2 = sdk.dart.dsend(reversed, "toList", []);
             const map = sdk.dart.dsend(list2, "asMap", []);
             const linkedMap = sdk.dart.dsend(sdk.collection.LinkedHashMap, "from", [map]);