+++ dwds/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.0-dev
+
+**Breaking Changes:**
+- Now require a `LoadStrategy` to `Dwds.start`. This package defines two
+  compatible load strategies, `RequireStrategy` and `LegacyStrategy.
+- `Dwds.start` function signature has been changed to accept one more parameter
+  of new interface type `ExpressionCompiler` to support expression
+  evaluation
+
 ## 1.0.1
 
 - Make the `root` optional for the `ProxyServerAssetReader`.
@@ -21,9 +30,6 @@
   clearly defined. The new abstraction is now consumed through `dwds.dart`.
 - `BuildRunnerAssetHandler` has been renamed to `ProxyServerAssetReader` and is
   now consumed through `dwds.dart`.
-- `Dwds.start` function signature has been changed to accept one more parameter
-  of new interface type `ExpressionCompiler` to support expression
-  evaluation
 
 ## 0.9.0